refactor backToProductListBannerButton
