+++ lib/screens/contact_us_page.dart
@@ -1,7 +1,7 @@
 import 'package:badiup/colors.dart';
 import 'package:badiup/screens/privacy_policy_page.dart';
 import 'package:badiup/sign_in.dart';
-import 'package:badiup/widgets/back_to_product_list_banner_button.dart';
+import 'package:badiup/widgets/banner_button.dart';
 import 'package:badiup/widgets/cart_button.dart';
 import 'package:flutter/material.dart';
 
@@ -17,7 +17,7 @@ class ContactUsPage extends StatelessWidget {
     return Scaffold(
       appBar: _buildAppBar(context),
       body: _buildBody(context),
-      bottomNavigationBar: BackToProductListBannerButton(),
+      bottomNavigationBar: BackToProductListBannerButton(context: context),
     );
   }
