+++ lib/widgets/banner_button.dart
@@ -1,4 +1,7 @@
 import 'package:badiup/colors.dart';
+import 'package:badiup/screens/admin_product_listing_page.dart';
+import 'package:badiup/screens/customer_home_page.dart';
+import 'package:badiup/sign_in.dart';
 import 'package:flutter/material.dart';
 
 class BannerButton extends StatelessWidget {
@@ -35,4 +38,24 @@ class BannerButton extends StatelessWidget {
       ],
     );
   }
+}
+
+class BackToProductListBannerButton extends BannerButton {
+  BackToProductListBannerButton({
+    Key key,
+    BuildContext context,
+  }) : super(
+          key: key,
+          text: '商品リスト',
+          onTap: () {
+            Navigator.pushReplacement(
+              context,
+              MaterialPageRoute(
+                builder: (context) => currentSignedInUser.isAdmin()
+                    ? AdminProductListingPage()
+                    : CustomerHomePage(),
+              ),
+            );
+          },
+        );
 }
